requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.0.8
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
idna==2.10
oauthlib==3.1.0
pillow==9.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.26.2
django-cors-headers==3.11.0